Move topn to core module.
